requirements.txt
openai==1.14.2
pdfplumber
simplejson
requests
PyYAML
pillow
reportlab
pandas
loguru
gradio
zhipuai
google-generativeai